// packages/b2c-tooling-sdk/src/clients/scapi-merchant-roles.generated.ts
export interface components {
    schemas: {
// ...
        RoleModulePermission: {
            name: string;
            type: string;
            application: string;
            system?: boolean;
            value?: string;
            values?: {
                [key: string]: string;
            };
        };
        RoleModulePermissions: {
            organization?: components["schemas"]["RoleModulePermission"][];
            site?: components["schemas"]["RoleModulePermission"][];
        };
        RoleFunctionalPermission: {
            name: string;
            type: string;
            value?: string;
            values?: {
                [key: string]: string;
            };
        };
// ...
        RoleLocalePermission: {
            localeId: components["schemas"]["LocaleCode"];
            type: string;
            value?: string;
            values?: {
                [key: string]: string;
            };
        };
        RoleLocalePermissions: {
            unscoped?: components["schemas"]["RoleLocalePermission"][];
        };
        RoleWebdavPermission: {
            folder: string;
            type: string;
            value?: string;
            values?: {
                [key: string]: string;
            };
        };
// ...
    };
    responses: never;
    parameters: {
        organizationId: components["schemas"]["OrganizationId"];
        expand: ("users" | "permissions")[];
        select: components["schemas"]["Select"];
        roleId: string;
        login: string;
    };
    requestBodies: never;
    headers: never;
    pathItems: never;
}